+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 - Use revamped parseTypeTag function instead of StructTag.fromString()
 - Allow use of generics in parseTypeTag
 - Rename publishModuleTransaction to publishPackageTransaction and fix functionality accordingly
+- Added toString() for type tags, and reference placeholder type
 
 ## 0.0.2 (2023-10-25)